Fix comment typo 수정

+++ src/main/java/escaper/backend/repository/comment/CommentRepository.java
@@ -1,7 +1,7 @@
 package escaper.backend.repository.comment;
 
-import escaper.backend.entity.commnet.Comment;
+import escaper.backend.entity.comment.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer>, CommentRepositoryCustom {
+public interface CommentRepository extends JpaRepository<Comment, Long>, CommentRepositoryCustom {
 }
